backend/requirements.txt
flask==3.0.3
pydash==5.1.2
PyJWT==2.8.0
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.0
python-dotenv==1.0.1
